(BSR)[BO] fix: issues reported by mypy-1.1.1 in backoffice v2/v3

+++ api/src/pcapi/routes/backoffice_v3/serialization/venues.py
@@ -16,7 +16,7 @@ class VenueOffersStats(OffersStats):
 
 class VenueStats(BaseModel):
     stats: VenueOffersStats
-    total_revenue: int
+    total_revenue: float
 
 
 class VenueDmsStats(BaseModel):
